update charts to point to correct location

+++ tools/filter-crds.sh
@@ -2,12 +2,15 @@
 
 for file in "$@"
 do
+  if [ ! -d $file ]; then
+    exit
+  fi
   echo "Filtering CRD's from $file"
   tempfile="$file.tmp"
   tempdir="tempdir"
   rm -rf $tempdir
   mkdir $tempdir
-  
+
   tar -xf $file -C $tempdir
   rm -rf $tempdir/**/crds
   cd $tempdir
